+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Small bug fixes and usability improvements
 - Renamed options for `dist.type` in `estimateExpressionShiftMagnitudes` to match the paper figure
 - Renamed inner `estimateExpressionShiftMagnitudes` to `estimateExpressionChange` (CRAN requirement)
+- Renamed `p.adj.cutoff` to `p.adj` in `estimateDEStabilityPerGene` for consistency with other functions